--- src/main/java/com/revealprecision/revealserver/api/v1/facade/dto/factory/LoginResponseFactory.java
@@ -1,7 +1,0 @@
-package com.revealprecision.revealserver.api.v1.facade.dto.factory;
-
-
-public class LoginResponseFactory {
-
-
-}
