+++ src/main/java/com/revealprecision/revealserver/api/v1/facade/factory/LoginResponseFactory.java
@@ -1,0 +1,7 @@
+package com.revealprecision.revealserver.api.v1.facade.factory;
+
+
+public class LoginResponseFactory {
+
+
+}
