+++ src/main/java/com/revealprecision/revealserver/api/v1/facade/models/ContactPoint.java
@@ -1,4 +1,4 @@
-package com.revealprecision.revealserver.api.v1.facade.dto.response;
+package com.revealprecision.revealserver.api.v1.facade.models;
 
 
 import java.io.Serializable;
