+++ src/main/java/com/revealprecision/revealserver/api/v1/facade/models/UserAssignmentResponse.java
@@ -1,4 +1,4 @@
-package com.revealprecision.revealserver.api.v1.facade.dto.response;
+package com.revealprecision.revealserver.api.v1.facade.models;
 
 import java.util.Set;
 import java.util.UUID;
